Switch `total_consistency_lock` to use `SendableRwLock`
Unfortunately the guard types of `std::sync::RwLock` are not `Send` and
therefore cannot be held over `.await` boundaries in multi-threaded
async runtime environments. This is in particular painful in case of
`ChannelManager`'s `total_consistency_lock` as it prohibits holding the
lock during `ChannelManager::process_pending_events_async`.

This implements `SendableRwLock`, a new readers/writer lock that can be
used as a drop-in replacement for `total_consistency_lock`. In order not
to unecessarily complicate the logic and avoid `unsafe` code, the lock
currently doesn't actually hold any data.

The implementation follows the algorithm given in
Buttlar, Dick, Jacqueline Farrell, and Bradford Nichols. Pthreads
programming: A POSIX standard for better multiprocessing. O'Reilly,
1996, pp. 86-89.