Move `channel_reestablish` messages to the `simple` message fuzzer
In 16d0f2ff4a4dd968f30e1fbbf938a332267ddcf2 the
`ChannelReestablish` messages was converted to the
`impl_writeable_msg` serialization macro which handles a TLV stream
suffix. In the case of our fuzzers, this implies we need to use the
`test_msg_simple` checker rather than the `test_msg` one.

`test_msg` checks that any bytes which were read must be present
when we write the message back out, which is useful for gossip
processing, however if we don't store ignored TLVs we cannot
possibly meet the requirements.

Instead, `test_msg_simple` only ensures that, if we serialized the
message we can round-trip it exactly, i.e. at least the fields we
understand round-trip.