Account for inflight htlcs in set_max_liquidity_msat
Before it was subtracting the amount from capacity_msat however, in all
the other functions, it is subtracted from available_capacity() to take
into account inflight htlcs. I believe this is a bug fix.

I also changed to used saturating_sub() instead of
checked_sub(..).unwrap_or(0) to be a bit cleaner.